changed dates
Hi Lindblom,

This week, we realized we have a scheduling conflict with the 8th Grade Selective Enrollment Exam on October 26th and Colloquium.  Yadi and I went through a number of possible solutions to have the 8th grade take this exam on the Colloquium day, but we could not find a path for the exam (CPS mandated, 1-day window) and Colloquium to coincide without significantly disrupting the schedules, spaces, and programs of a significant number of Colloquium courses.

So, today we consulted with the PPC, the PPLC, the Colloquium Committee, Counseling, and Admin, and we've decided to move the Colloquium day on the last week of October to Friday, October 28th.

I know a month out is tight, and I apologize for not catching the conflict before.  This mistake is on me.

The last week of October will now look like this:

Monday, October 24th: 1-4 day
Tuesday, October 25th: 5-8 day
Wednesday, October 26th: 1-4 day (but no bells due to 8th grade testing)
Thursday, October 27th: 5-8 day
Friday, October 28th: Colloquium day

If you already have a field trip scheduled for October 26th, the field trip can still run (unless 8th graders are involved).  We've been running field trips all school year on non-colloquium days, and it has been fine.

If you need help communicating with an outside partner about the schedule change, let me know.  Again, I apologize for the inconvenience, and I'm happy to support you if you need help making this change work.

Niki